Everywhere: Started work on upgrading to version 89
This first badge of changes do not pass the build, however make it quite far.

The CI pipeline got adjusted to allow for a longer build, the time needed for the previous
steps strongly suggests the need of this change. If it was really needed can only
be checked by a passing build however

The package build definitions have also been updated, setting the new chromium
and ungoogled-chromium version as well as applying different patches for libstdc++
build support

The subpixel-anti-aliasing-in-FreeType-2.8.1 patch has been removed as it now matches upstream source
files