CI+README: Upgrade to container based building system
This commit introduces a rework of the continous integration pipeline
used to test build on every push and pull request.

Actual compiling has been moved into a container that gets built in
the first step and is then reused as a base for all following steps.
The image of the container, also being published in the GitHub Package
Registry for users as a self build option contains all nessearry
dependencies for building as well as all base files, currently including
patches, chromium sources and the chromium lanucher wrapper.
The previously hard coded makepkg.conf has been removed in order to
replicate a normal arch install more closely and catch related bugs.

The workflow is split into currently 6 jobs all running the same steps
to circumvent the 6 hours job timeout set by GitHub Actions.
Compared to the previous iteration of the workflow, building steps
will get skipped properly in case a package has already been compiled.

Besides the compiling of the package, the test workflow now also
runs a simple test to determent if the chromium executable can be
installed and run correctly.