CI: Replace building stage with TypeScript Action
This extracts logic to control the build container
in a stage to a local Action written in TypeScript.

Compared to the previous YAML steps this is a lot
more scalable, easier to maintain and allows for
more complicated logic.

Additionally, as pull requests do not have permission
to push to the container registry, the container image
now gets distributed to each stage using an artifact if
required.