CI: Publish AUR archive alongside releases
This patch contains initial work towards releasing an
ungoogled-chromium-bin AUR package.

A previous versions of the package was removed for following reason:
"the AUR is not intended to provide an index for PKGBUILDs that
repackage the build package originally created by another PKGBUILD".

Asking on the aur-general mailing list about the best way to submit
a package regardless concluded in creating a seperate archive not
created by makepkg directly (stated as being the best option by
Alexander Epaneshnikov <alex19ep@archlinux.org>: https://lists.archlinux.org/pipermail/aur-general/2021-September/036583.html)

The release workflow has been modified to create a seperate archive
called "aur.tar.zst" for every release, which does not include makepkg
metadata. I don't know how to better comply with not repackaging an
existing package archive as the build process is based on makepkg and
rewriting a build system from scratch is not feasible.

Currently pushing updates to the AUR automatically is commented out in
the workflow, a future commit will update that section. As such the
newly added PKGBUILD file for submitting to the AUR is also not used.
I intend to use the archive created by CI to manually submit the initial
package version.

Work towards #143