CI: Attempt to recompress tarball before copying it into the container
As the chromium source tarball converges into a black hole,
we have to save disk space by throwing away useless files.

This patch attempts to do so by streaming a download and
extract with filters to recompress the tarball with zstd.
The hope is that the runner will have just enough space
for extracting the tarball just before building a chunky
OCI image.